+++ library/std/src/os/fortanix_sgx/io.rs
