--- library/std/src/sys/sgx/ext/mod.rs
@@ -1,5 +1,0 @@
-#![unstable(feature = "sgx_platform", issue = "56975")]
-
-pub mod arch;
-pub mod ffi;
-pub mod io;
